+++ pyas2/__init__.py
@@ -1,5 +1,5 @@
 # Set the version
-__version__ = '1.0.1'
+__version__ = '1.0.2'
 
 default_app_config = 'pyas2.apps.Pyas2Config'